fix: address PR review feedback for native Win32 tool handlers
- Fix chord handling: use send_chord() instead of looping send_key()
  so modifier combinations (Ctrl+C etc.) work correctly
- Fix parse_send_keys: remove reversed modifiers from output since
  send_chord handles release automatically
- Fix apps schema: maintain backward-compatible 'apps' field alongside
  the new 'windows' detail, filter to titled windows only
- Fix text response: report char count instead of byte length
- Fix scroll: remove redundant i32 cast (field is already i32)
- Add chord.rs: batched SendInput for proper modifier+key combos

+++ src/tool/computer_use/platform/windows/input/scroll.rs
@@ -6,7 +6,7 @@ use crate::tool::computer_use::input::ComputerUseInput;
 /// Send scroll wheel event using native Win32 API.
 pub async fn handle_scroll(input: &ComputerUseInput) -> anyhow::Result<crate::tool::ToolResult> {
     let amount = input.scroll_amount.unwrap_or(-120);
-    send_scroll(amount as i32)?;
+    send_scroll(amount)?;
     Ok(super::super::response::success_result(serde_json::json!({
         "scrolled": amount
     })))
